network: Use wicked to control bond slaves (bsc#1035127)
Currently, when adding slaves to a bond, the interface will be brought
down and then back up. If wicked is not told about this state of
affairs, wicked will add the configured IP address back onto the
interface when it is bought back up.

To counteract this, we run wicked ifdown for any slave interface, as
well as no longer bringing the slave back up immediately and then run
wicked ifup for every interface if the template files have changed, so
it isn't run every time chef fires.