Use "wicked ifup all" instead of separate calls per interface
This makes sure that the "wicked ifup" call happens after all templated
for the current run were updated. Also wicked will make sure to process the
interfaces in the right order.

Note, this partially brings back what we reverted with
41e310cf21642cf0dc2f2fa0b7294c1db7821b59, but calling "ifup" instead of
"ifreload" has the advantage that will not take the interfaces down
temporary. It will also keep all IP addresses that are not managed by
wicked (i.e. the pacemaker managed cluster VIPs). So we should avoid the
issues that the "ifreload" patch created.